Rewrite unit tests under base directory
